Config float fields: switch f32 → f64 to round-trip cleanly through TOML
Reported: editing `[ai] threshold = 0.1` via the new Config tab and
saving rewrote it as `0.10000000149011612`. Same applies to every
float-valued config field.

Cause: those fields were typed as `f32`. The save path goes
JSON (f64) → serde → CatalogConfig (f32) → toml::to_string_pretty.
The cast to f32 at deserialise time loses precision (0.1 has no
exact f32 representation; closest is ~0.1000000015), and
toml::to_string_pretty re-widens the f32 to f64 before printing,
exposing that imprecision as digits in the output.

Fix: use f64 across all float-typed config fields (TOML's native
float type). 0.1 has no exact representation as f64 either, but
TOML's serializer uses ryu's shortest-roundtrip formatter, which
prints "0.1" for the f64 value parsed from "0.1". Cast to f32 at
the boundary where downstream consumers actually want f32 (LLM API,
ONNX inference, contact-sheet renderer).

Touched fields:
  ai.threshold, ai.face_cluster_threshold, ai.face_min_confidence
  vlm.temperature, vlm.top_p, vlm.repeat_penalty
  vlm.model_config.<name>.{temperature,top_p,repeat_penalty}
  contact_sheet.margin

Boundary casts added at:
  - config::VlmConfig::params_for_model — produces VlmParams (f32)
  - web::routes::ai::tags batch_auto_tag, classify_asset
  - web::routes::ai::faces batch_detect_faces, detect_asset_faces,
    auto_cluster
  - commands::run_faces_command — CLI flag default
  - commands::run_auto_tag_command — CLI flag default
  - commands::run_contact_sheet_command — CLI flag default

Regression test `float_values_round_trip_cleanly`: parse `[ai]
threshold = 0.2`, serialise back, assert the literal `threshold =
0.2` appears in the output and no f32-imprecision long form
(`0.20000000…`) leaks in. Used 0.2 rather than 0.1 so the [ai]
section actually serialises (0.1 is the default, gets skipped via
`skip_serializing_if`).

Tests: 792 / 899 lib (+1 each), 249 / 276 cli unchanged.

+++ src/commands.rs
@@ -206,8 +206,11 @@ pub fn run_faces_command(
             let _ = maki::face_store::FaceStore::initialize(catalog.conn());
             let face_store = maki::face_store::FaceStore::new(catalog.conn());
 
-            let thresh = threshold.unwrap_or(config.ai.face_cluster_threshold);
-            let min_conf = min_confidence.unwrap_or(config.ai.face_min_confidence);
+            // CLI threshold/min_confidence args are f32 (clap default for `--threshold`);
+            // config carries f64 since v4.5.5 (for clean TOML round-trips). Cast to f32
+            // at this boundary so the rest of the function sees a single type.
+            let thresh = threshold.unwrap_or(config.ai.face_cluster_threshold as f32);
+            let min_conf = min_confidence.unwrap_or(config.ai.face_min_confidence as f32);
 
             // Resolve scope to asset IDs (same pattern as maki embed)
             let scoped_ids: Option<Vec<String>> = if query.is_some() || asset.is_some() || volume.is_some() {
@@ -1562,7 +1565,8 @@ pub fn run_auto_tag_command(
         );
     }
 
-    let threshold = threshold.unwrap_or(config.ai.threshold);
+    // CLI threshold is f32; config field is f64 since v4.5.5. Cast at the boundary.
+    let threshold = threshold.unwrap_or(config.ai.threshold as f32);
 
     // Resolve labels
     let label_list: Vec<String> = if let Some(ref labels_path) = labels {
@@ -3827,7 +3831,9 @@ pub fn run_contact_sheet_command(
         .parse()?;
 
     let cs_quality = quality.unwrap_or(cs_defaults.quality);
-    let cs_margin = margin.unwrap_or(cs_defaults.margin);
+    // CLI --margin is f32; config carries f64 since v4.5.5. Cast at the boundary;
+    // ContactSheetConfig.margin_mm downstream is f32.
+    let cs_margin = margin.unwrap_or(cs_defaults.margin as f32);
 
     let cs_config = ContactSheetConfig {
         layout: cs_layout,
